+++ src/app/page.tsx
@@ -18,7 +18,8 @@ import Spinner from "@/components/common/Spinner";
 import GetStarted from "@/components/features/Questions/GetStarted";
 import Result from "@/components/features/Questions/Results";
 import { useQuestionStore } from "@/store/questions";
-import QuestionWrappers from "@/components/wrappers/Question";
+import QuestionWrappers from "@/components/wrappers/question/Question";
+import MainWrappers from "@/components/wrappers/question/Main";
 
 export interface QuestionData {
   id: string;
@@ -64,7 +65,7 @@ const Reading = () => {
   }
 
   return (
-    <div className="flex flex-col md:flex-row h-screen">
+    <MainWrappers>
       <Sidebar
         title="Reading"
         svg={<BookSVG />}
@@ -107,7 +108,7 @@ const Reading = () => {
       </QuestionWrappers>
 
       {isAnnouncerModalOpen && <StreakAnnouncer />}
-    </div>
+    </MainWrappers>
   );
 };
