+++ src/types/wrapper.ts
@@ -1,0 +1,4 @@
+export interface Wrapper {
+    children: React.ReactNode; // Correct spelling here
+  }
+
